+++ changelog.md
@@ -5,6 +5,8 @@
 - Added random seed parameters to each of the sampling functions in both the C# and Rust code. These random seeds can be used to deterministically recreate sampled points.
   - Added `autoSeed` parameter to PincushionManager. Set this to false to manually seed each sampled mesh, thereby allowing for deterministic outcomes.
   - The Rust code now uses `fastrand` to generate random numbers instead of `rand`. This is, as implied, somewhat faster, and is portable (results should be the same regardless of OS or hardware).
+- Added: `PincushionRenderer.GetTransformedPoints()`. Returns the sampled points, transformed by the transform matrix.
+  - Rust: Added `transform_points(matrix, points)`. This is *probably* faster than transforming the points in C#.
 - Moved all code in `ffi.rs` into `lib.rs`
 - Fixed: The binary used to generate C# native bindings has an invalid target file path.